requirements.txt
django
celery
djangorestframework
django-celery-beat
redis
django-redis
aiohttp